+++ tests/PdfCompareBasicTest.php
@@ -91,7 +91,7 @@ class PdfCompareBasicTest extends AbstractCompareTestCase
         $doc->addPage();
         $doc->cell(null, 5.0, 'Greek: Γειά σου κόσμος', PdfBorder::all(), PdfMove::RIGHT, PdfTextAlignment::RIGHT, true);
 
-        $doc->setDisplayMode(PdfZoom::FULL_PAGE, PdfLayout::SINGLE);
+        $doc->setDisplayMode(PdfZoom::FULL_PAGE, PdfLayout::SINGLE_PAGE);
 
         $doc->multiCell(text: self::COMMENT);
     }